fix: move generateUniqueMeetingId to db layer, deduplicate formatDateAsMeetingId
generateUniqueMeetingId queries the database, so it belongs in
src/lib/db/meetings.ts, not src/lib/utils/. Moving it out makes
meetingId.ts a pure utility with no Prisma import, allowing the
client-side AddMeetingForm to import formatDateAsMeetingId directly
instead of maintaining a duplicate copy.

The client copy was missing timeZone: 'Europe/Athens', which caused
the ID preview to diverge from the server-generated ID for users in
other timezones.
